+++ mne_bids/utils.py
@@ -67,7 +67,7 @@ def _get_ch_type_mapping(fro='mne', to='bids'):
     if fro == 'mne' and to == 'bids':
         mapping = dict(eeg='EEG', misc='MISC', stim='TRIG', emg='EMG',
                        ecog='ECOG', seeg='SEEG', eog='EOG', ecg='ECG',
-                       resp='RESP', bio='MISC',
+                       resp='RESP', bio='MISC', dbs='DBS',
                        # MEG channels
                        meggradaxial='MEGGRADAXIAL', megmag='MEGMAG',
                        megrefgradaxial='MEGREFGRADAXIAL',
@@ -80,7 +80,7 @@ def _get_ch_type_mapping(fro='mne', to='bids'):
                        RESP='resp',
                        # No MEG channels for now
                        # Many to one mapping
-                       VEOG='eog', HEOG='eog')
+                       VEOG='eog', HEOG='eog', DBS='dbs')
     else:
         raise ValueError('Only two types of mappings are currently supported: '
                          'from mne to bids, or from bids to mne. However, '
@@ -125,7 +125,7 @@ def _handle_datatype(raw, datatype, verbose=True):
             datatype = 'meg'
     else:
         datatypes = list()
-        ieeg_types = ['seeg', 'ecog']
+        ieeg_types = ['seeg', 'ecog', 'dbs']
         if any(ieeg_type in raw for ieeg_type in ieeg_types):
             datatypes.append('ieeg')
         if 'meg' in raw:
@@ -270,7 +270,6 @@ def _infer_eeg_placement_scheme(raw):
         return placement_scheme
 
     # How many of the channels in raw are based on the extended 10/20 system
-    raw.load_data()
     sel = pick_types(raw.info, meg=False, eeg=True)
     ch_names = [raw.ch_names[i] for i in sel]
     channel_names = [ch.lower() for ch in ch_names]
@@ -459,7 +458,7 @@ def _check_datatype(raw, datatype):
     elif datatype == 'meg' and datatype in raw:
         datatype_matches = True
     elif datatype == 'ieeg':
-        ieeg_types = ('seeg', 'ecog')
+        ieeg_types = ('seeg', 'ecog', 'dbs')
         if any(ieeg_type in raw for ieeg_type in ieeg_types):
             datatype_matches = True
     if not datatype_matches: